+++ main.go
@@ -1,0 +1,10 @@
+package uilive
+
+import (
+	"github.com/Adaendra/uilive/pkg/writer"
+)
+
+// New returns a new Writer with defaults
+func New() *writer.Writer {
+	return writer.New()
+}