+++ pkg/writer/writer.go
@@ -1,8 +1,9 @@
-package uilive
+package writer
 
 import (
 	"bytes"
 	"errors"
+	"github.com/Adaendra/uilive/pkg/terminal"
 	"io"
 	"os"
 	"sync"
@@ -11,6 +12,10 @@ import (
 
 // ESC is the ASCII code for escape character
 const ESC = 27
+
+// ASCII_END_COLOR_CODE is the last character in an ASCII escape code that
+// changes the color
+const ASCII_END_COLOR_CODE = 'm'
 
 // RefreshInterval is the default refresh interval to update the ui
 var RefreshInterval = time.Millisecond
@@ -57,7 +62,7 @@ type newline struct {
 
 // New returns a new Writer with defaults
 func New() *Writer {
-	termWidth, _ = getTermSize()
+	termWidth, _ = terminal.GetTermSize()
 	if termWidth != 0 {
 		overFlowHandled = true
 	}
@@ -85,15 +90,30 @@ func (w *Writer) Flush() error {
 
 	lines := 0
 	var currentLine bytes.Buffer
-	for _, b := range w.buf.Bytes() {
-		if b == '\n' {
+	var escaping bool
+
+	bytes := w.buf.Bytes()
+	bytesLen := len(bytes)
+	for i, b := range bytes {
+		switch {
+		case b == '\n':
 			lines++
 			currentLine.Reset()
-		} else {
-			currentLine.Write([]byte{b})
-			if overFlowHandled && currentLine.Len() > termWidth {
+		case b == ESC:
+			// All ASCII escape sequences start with ESC, followed by a second
+			// byte in the 0x40 to 0x5F range.
+			if i < bytesLen-1 && bytes[i+1] >= 0x40 && bytes[i+1] <= 0x5F {
+				escaping = true
+			}
+		case escaping && b == ASCII_END_COLOR_CODE:
+			escaping = false
+		default:
+			if overFlowHandled && currentLine.Len() >= termWidth {
 				lines++
 				currentLine.Reset()
+			}
+			if !escaping {
+				currentLine.Write([]byte{b})
 			}
 		}
 	}